requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
pandas
openpyxl
requests
xlrd
pytest
python-dotenv
apscheduler